WIP: rewrite to use more generic RPC
problems:
right now sending things blindly just doesn't work. the ui has no
idea if it's interested in an object before it sees the registration, by which
point it has lost the data. and that's top level objects: children are
even worse... OBJECT_CREATE on a child is meaningless until it knows the
child/parent relationship...

theory:
OBJECT_SUBSCRIBE replacing OBJECT_REGISTER. no OBJECT_CREATE can be sent
before it has been subscribed to.

    ui: subscribe foo
    backend: object foo <uuid> <data>
    ui: unsubscribe foo

a list model could then be like:

    ui: subscribe people
    backend: object people <uuid> <data>
        (where 'data' is an array of JS objects: uuid & actual data)
    invoke people_uuid append <uuid> <data>
    invoke people_uuid remove <uuid>
    etc.

Remote object handler virtuals:

    objectFound(uuid, initialData)
    objectChanged(uuid, newData)
    methodInvoked(method, data)
    objectLost()

So a model would create a remote object handler with a given name, and
the right virtuals. That would send a subscribe, which would in turn
call virtuals as messages arrived..