Introduce a root object on the connection
As the global namespace of objects goes away, the only way to get an
object is as a value from another object. To solve the chicken and egg
problem, this introduces the root object.

There is exactly one root object on the connection. It is always
available as QBackendConnection's root property. It uniquely has an
identifier of "root", but this is an implementation detail. The type and
format of the root object is entirely controlled by the backend
implementation -- the important detail is that it's always available.

Effectively, think of this as an equivalent to QML's root context.