Replace protocol with something more JSON-based
This little protocol was due for a refactor to behave more consistently,
and some upcoming work needs the handling of different messages to be a
bit more consistent.

All commands and behavior remains the same, but commands are now JSON
objects, framed with just a (ASCII) size and a newline. One benefit to
this is that the entire command can be read without command-specific
logic.