Add a blocking QBackendConnection::waitForMessage method
Hear me out.

For the "seamless objects" goal, properties have to be readable in
normal ways, which means the value needs to be immediately available. We
keep all values for an object locally, so that's fine, until we get to
objects that reference other objects.

To avoid having to copy the entire tree of every object that is referred
to by any object over ahead of time, we need a way to fill object data
in on-demand. For that to be possible, we have to be able to query it
synchronously.

This shouldn't be overused, but for these particular cases it's a
critical functionality for having a reasonable API.