Start generating metaobjects for QBackendObject
QBackendObject now takes a JSON type description at construction, and
builds a QMetaObject out of this to transparently mirror the remote
object's properties, methods, and signals.

For the moment, the metaObject isn't used, so behavior is unchanged.
Next steps are to send type info from the backend and make sure it's
handled properly, then to fully transform QBackendObject.