Implement a seamless Go object API
This still wants for detailed documentation, but briefly:

The old API involved explicitly creating a *Store, from the connection,
for every object that would be published. Those ended up being
constantly annoying to store and create; it was intrusive to everything.

Now that all objects exist somewhere under a root object (whether as
fields, members in arrays, parameters in signals), there's an
opportunity to initialize any object before it's used. We can take
advantage of this, along with a series of API tricks and a shitton of
reflection, for zero-initialization straightforward use of Go objects.

Effectively, objects are declared simply by embedding the QObject type
into any struct:

type Thing struct {
    qbackend.QObject

    Word     string
    SayHello func(string) `qbackend:"word"`
}

func (t *Thing) SomeMethod() {
}

Other than its internal use, this adds a handful of useful methods
directly to the struct, e.g. Emit().

All exported methods of this struct will be automatically available as
methods in QML. All exported fields, excluding those below, are
available as properties.

Fields with a func() type are signals. Their parameters are passed as
expected, but they must be tagged with the names of parameters for QML.

When this object is encountered while marshaling any other data (which
is the only way it could be known), it's initialized transparently,
assigning a unique ID, reflecting the type, and creating signals.

At initialization, any signals that are still nil will be assigned to
emit their signal when called. So, signals can be emitted with simply a
`thing.SayHello("hello")`.