Rearrange QBackendConnection handling of objects
The disconnect between m_subscribedObjects, m_objects, and reality was
becoming a problem.

The QBackendRemoteObject API has been changed and cleaned up a little;
it's still implemented by QBackendObjectPrivate. QBackendConnection
keeps a mapping of identifiers to QBackendRemoteObjects, and the QBRO
has a method to retrieve the "front" QObject.

Objects are responsible for adding and removing themselves from the
conenction on (con|de)struction.