Keep a count of references between objects
This is a first step towards being able to release unreachable objects
and allow them to be garbage collected, which is currently never
possible because of the ID map, and because we can't know if they will
be used by the client again.

Here, objects will keep track of any other objects that are referenced
within their properties -- this is easy to do, because initObjectsUnder
already does this scan to initialize any new QObject. This is used to
keep a refCount on each objectImpl for the total references from other
objects.

If an objectImpl's refCount (by other objects) is 0 and Ref (by client)
is false, there is only one remaining way that an object could be
referenced: as a direct parameter to a signal. That is the next case to
resolve.