Change protocol around adding/removing/resetting objects
The client is now expected to OBJECT_REF any object that it is (or might
be) interested in, OBJECT_DEREF when no longer interested, and can use
OBJECT_QUERY to ask for data, which arrives as OBJECT_RESET.

In practice, this means the backend can track which objects a client has
a BackendObjectPrivate for, even if they haven't been used yet. Because
that will be the full set of objects the client can access without
reading new properties, the backend can be sure of when an object is
safely unreferenced.