Autoconfigure QBackendConnection when possible
Especially when used as a singleton, QBackendConnection will try to find
the backend URL from the QML context (qbackendUrl), commandline
(-qbackend), or environment (QBACKEND_URL), in that order.

This means there is nothing required to start using a connection from
QML; everything happens automatically when you reference the root
object.