Expose the root object as singleton rather than connection
There is no connection API that is particularly relevant when using the
singleton type, so this can become one little step more convenient by
exposing the root object itself as the singleton. This requires a bit of
creativity to get the object type early enough, but it's not too
different from how it would work otherwise.

In practice, if your backend has a property "whoo" in the root object,
it can be used with nothing more than:

import QBackend 1.0
...
    text: Backend.whoo