Populate models with just-in-time batched requests
The backend model can be configured with a batch size. When set, the
internal model signals will only send data for the first batchSize rows,
with a count for the remaining rows. This applies to reset as well, so
initialization of the model will only load batchSize rows of data.

If the model needs data for a row it doesn't have, it will be
synchronously requested as a batch surrounding that row index.

For the moment, batch size is hardcoded in the plugin as 100, and the
cache is kept around 1000 entries.