Omit type descriptions the client has already seen
The full description of a type description is only needed when first
building a metaobject; for all future objects, the type is cached.
Taking advantage of that, the size of object refs can be dramatically
reduced by omitting everything but the type name for types that the
client has acknowledged seeing.

After the client has referenced an object, that type is marked as known
and the description will be omitted thereafter.