Call a componentDestruction method for instantiated objects
This is equivalent to Component.onDestruction in QML, and has the same
behavior as componentComplete -- if the backend instantiable type has a
method with the right signature, it will be called automatically.

This is useful for freeing backend resources on instantiated objects,
or removing internal references. It is not necessary for the basic
garbage collection of all qbackend objects.