Avoid issues with reusing gabage collected objects
If a QJSValue-wrapped QObject is garbage collected, it's marked as
deleted immediately but may actually be deleted later. This left a case
where a QBackendObject could be garbage collected and then reused within
the same event loop iteration. QML would give a null QJSValue for that
case.

This is now detected and the object will get a new instance immediately
to replace the soon-to-be-deleted one.