Simplify connection states further
There is not actually any need for ROOT to be part of the initial
connection state. Once there is a QML engine, any sort of message can be
handled, and if ROOT is needed it will be handled as a blocking message
like any other. This simplifies the states during connection quite a
bit.

Also, clean up the handling of synchronous messages a bit to fix issues
with recursive calls.